chore: verify zero osutk opentk refs requirement
Agent-Logs-Url: https://github.com/winnerspiros/osu-framework/sessions/10551308-4e73-431f-84eb-ef77b3006f1c

+++ osu.Framework/Graphics/OpenGL/Buffers/GLLinearBuffer.cs
@@ -50,7 +50,7 @@ namespace osu.Framework.Graphics.OpenGL.Buffers
                 for (int i = 0; i < amountVertices; i++)
                     indices[i] = (ushort)i;
 
-                GL.BufferData(BufferTarget.ElementArrayBuffer, (IntPtr)(amountVertices * sizeof(ushort)), indices, BufferUsageHint.StaticDraw);
+                GL.BufferData(BufferTarget.ElementArrayBuffer, amountVertices * sizeof(ushort), indices, BufferUsageHint.StaticDraw);
 
                 GLLinearIndexData.MaxAmountIndices = amountVertices;
             }
